(#6) refac : 시간 함수모듈화(목록,상세페이지 둘다쓰일예정)
